requirements.txt
# PyTorch - install with CUDA for your GPU:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# Or CPU only:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0

# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
tqdm>=4.65.0
Pillow>=10.0.0
pyyaml>=6.0
faiss-cpu>=1.7.4
